Refactor collect_dataset page into folder
